requirements.txt
flask
python-telegram-bot==13.7
python-binance==1.0.17
